sivia/requirements.txt
Flask>=2.0
flask-cors
python-dotenv
google-generativeai
requests
beautifulsoup4
